+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Fix detection of Steam library folders using non-standard capitalizations for `steamapps`
+
 ## [1.10.2] - 2023-02-13
 ### Added
  - Launch application with fixed locale settings if Steam Deck is used and non-existent locales are configured